requirements.txt
spotipy
pandas
sqlalchemy
streamlit
dbt-core
dbt-sqlite 
python-dotenv
